Migrate to .hpp files
