+++ fuse_loss/include/fuse_loss/arctan_loss.hpp
@@ -31,8 +31,8 @@
  *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  *  POSSIBILITY OF SUCH DAMAGE.
  */
-#ifndef FUSE_LOSS_ARCTAN_LOSS_H
-#define FUSE_LOSS_ARCTAN_LOSS_H
+#ifndef FUSE_LOSS_ARCTAN_LOSS_HPP_
+#define FUSE_LOSS_ARCTAN_LOSS_HPP_
 
 #include <fuse_core/loss.hpp>
 
@@ -150,4 +150,4 @@ private:
 
 BOOST_CLASS_EXPORT_KEY(fuse_loss::ArctanLoss);
 
-#endif  // FUSE_LOSS_ARCTAN_LOSS_H
+#endif  // FUSE_LOSS_ARCTAN_LOSS_HPP_
