+++ fuse_loss/include/fuse_loss/cauchy_loss.hpp
@@ -31,8 +31,8 @@
  *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  *  POSSIBILITY OF SUCH DAMAGE.
  */
-#ifndef FUSE_LOSS_TOLERANT_LOSS_H
-#define FUSE_LOSS_TOLERANT_LOSS_H
+#ifndef FUSE_LOSS_CAUCHY_LOSS_HPP_
+#define FUSE_LOSS_CAUCHY_LOSS_HPP_
 
 #include <fuse_core/loss.hpp>
 
@@ -48,29 +48,28 @@ namespace fuse_loss
 {
 
 /**
- * @brief The TolerantLoss loss function.
+ * @brief The CauchyLoss loss function.
  *
- * This class encapsulates the ceres::TolerantLoss class, adding the ability to serialize it and load it dynamically.
+ * This class encapsulates the ceres::CauchyLoss class, adding the ability to serialize it and load it dynamically.
  *
  * See the Ceres documentation for more details. http://ceres-solver.org/nnls_modeling.html#lossfunction
  */
-class TolerantLoss : public fuse_core::Loss
+class CauchyLoss : public fuse_core::Loss
 {
 public:
-  FUSE_LOSS_DEFINITIONS(TolerantLoss)
+  FUSE_LOSS_DEFINITIONS(CauchyLoss)
 
   /**
    * @brief Constructor
    *
-   * @param[in] a TolerantLoss parameter 'a'. See Ceres documentation for more details
-   * @param[in] b TolerantLoss parameter 'b'. See Ceres documentation for more details
+   * @param[in] a CauchyLoss parameter 'a'. See Ceres documentation for more details
    */
-  explicit TolerantLoss(const double a = 1.0, const double b = 0.1);
+  explicit CauchyLoss(const double a = 1.0);
 
   /**
    * @brief Destructor
    */
-  ~TolerantLoss() override = default;
+  ~CauchyLoss() override = default;
 
   /**
    * @brief Perform any required post-construction initialization, such as reading from the parameter server.
@@ -118,16 +117,6 @@ void initialize(
   }
 
   /**
-   * @brief Parameter 'b' accessor.
-   *
-   * @return Parameter 'b'.
-   */
-  double b() const
-  {
-    return b_;
-  }
-
-  /**
    * @brief Parameter 'a' mutator.
    *
    * @param[in] a Parameter 'a'.
@@ -137,19 +126,8 @@ void initialize(
     a_ = a;
   }
 
-  /**
-   * @brief Parameter 'b' mutator.
-   *
-   * @param[in] b Parameter 'b'.
-   */
-  void b(const double b)
-  {
-    b_ = b;
-  }
-
 private:
-  double a_{ 1.0 };  //!< TolerantLoss parameter 'a'. See Ceres documentation for more details
-  double b_{ 0.1 };  //!< TolerantLoss parameter 'b'. See Ceres documentation for more details
+  double a_{ 1.0 };  //!< CauchyLoss parameter 'a'. See Ceres documentation for more details
 
   // Allow Boost Serialization access to private methods
   friend class boost::serialization::access;
@@ -165,12 +143,11 @@ private:
   {
     archive & boost::serialization::base_object<fuse_core::Loss>(*this);
     archive & a_;
-    archive & b_;
   }
 };
 
 }  // namespace fuse_loss
 
-BOOST_CLASS_EXPORT_KEY(fuse_loss::TolerantLoss);
+BOOST_CLASS_EXPORT_KEY(fuse_loss::CauchyLoss);
 
-#endif  // FUSE_LOSS_TOLERANT_LOSS_H
+#endif  // FUSE_LOSS_CAUCHY_LOSS_HPP_
