+++ fuse_loss/include/fuse_loss/scaled_loss.hpp
@@ -1,7 +1,7 @@
 /*
  * Software License Agreement (BSD License)
  *
- *  Copyright (c) 2019, Clearpath Robotics
+ *  Copyright (c) 2020, Clearpath Robotics
  *  All rights reserved.
  *
  *  Redistribution and use in source and binary forms, with or without
@@ -31,8 +31,8 @@
  *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  *  POSSIBILITY OF SUCH DAMAGE.
  */
-#ifndef FUSE_LOSS_SOFTLONE_LOSS_H
-#define FUSE_LOSS_SOFTLONE_LOSS_H
+#ifndef FUSE_LOSS_SCALED_LOSS_HPP_
+#define FUSE_LOSS_SCALED_LOSS_HPP_
 
 #include <fuse_core/loss.hpp>
 
@@ -40,6 +40,7 @@
 #include <boost/serialization/base_object.hpp>
 #include <boost/serialization/export.hpp>
 
+#include <memory>
 #include <ostream>
 #include <string>
 
@@ -48,28 +49,29 @@ namespace fuse_loss
 {
 
 /**
- * @brief The SoftLOneLoss loss function.
+ * @brief The ScaledLoss loss function.
  *
- * This class encapsulates the ceres::SoftLOneLoss class, adding the ability to serialize it and load it dynamically.
+ * This class encapsulates the ceres::ScaledLoss class, adding the ability to serialize it and load it dynamically.
  *
- * See the Ceres documentation for more details. http://ceres-solver.org/nnls_modeling.html#lossfunction
+ * See the Ceres documentation for more details: http://ceres-solver.org/nnls_modeling.html#lossfunction
  */
-class SoftLOneLoss : public fuse_core::Loss
+class ScaledLoss : public fuse_core::Loss
 {
 public:
-  FUSE_LOSS_DEFINITIONS(SoftLOneLoss)
+  FUSE_LOSS_DEFINITIONS(ScaledLoss)
 
   /**
    * @brief Constructor
    *
-   * @param[in] a SoftLOneLoss parameter 'a'. See Ceres documentation for more details
+   * @param[in] a ScaledLoss parameter 'a'. See Ceres documentation for more details.
+   * @param[in] loss The loss function to scale. Its output is scaled/multiplied by 'a'.
    */
-  explicit SoftLOneLoss(const double a = 1.0);
+  explicit ScaledLoss(const double a = 1.0, const std::shared_ptr<fuse_core::Loss>& loss = nullptr);
 
   /**
    * @brief Destructor
    */
-  ~SoftLOneLoss() override = default;
+  ~ScaledLoss() override = default;
 
   /**
    * @brief Perform any required post-construction initialization, such as reading from the parameter server.
@@ -95,7 +97,7 @@ void initialize(
   void print(std::ostream& stream = std::cout) const override;
 
   /**
-   * @brief Return a raw pointer to a ceres::LossFunction that implements the loss function
+   * @brief Return a raw pointer to a ceres::LossFunction that implements the loss function.
    *
    * The Ceres interface requires a raw pointer. Ceres will take ownership of the pointer and promises to properly
    * delete the loss function when it is done. Additionally, Fuse promises that the Loss object will outlive any
@@ -117,6 +119,16 @@ void initialize(
   }
 
   /**
+   * @brief Parameter 'loss' accessor.
+   *
+   * @return Parameter 'loss'.
+   */
+  std::shared_ptr<fuse_core::Loss> loss() const
+  {
+    return loss_;
+  }
+
+  /**
    * @brief Parameter 'a' mutator.
    *
    * @param[in] a Parameter 'a'.
@@ -126,8 +138,19 @@ void initialize(
     a_ = a;
   }
 
+  /**
+   * @brief Parameter 'loss' mutator.
+   *
+   * @param[in] loss Parameter 'loss'.
+   */
+  void loss(const std::shared_ptr<fuse_core::Loss>& loss)
+  {
+    loss_ = loss;
+  }
+
 private:
-  double a_{ 1.0 };  //!< SoftLOneLoss parameter 'a'. See Ceres documentation for more details
+  double a_{ 1.0 };  //!< ScaledLoss parameter 'a'. See Ceres documentation for more details
+  std::shared_ptr<fuse_core::Loss> loss_{ nullptr };  //!< The loss function to scale
 
   // Allow Boost Serialization access to private methods
   friend class boost::serialization::access;
@@ -143,11 +166,12 @@ private:
   {
     archive & boost::serialization::base_object<fuse_core::Loss>(*this);
     archive & a_;
+    archive & loss_;
   }
 };
 
 }  // namespace fuse_loss
 
-BOOST_CLASS_EXPORT_KEY(fuse_loss::SoftLOneLoss);
+BOOST_CLASS_EXPORT_KEY(fuse_loss::ScaledLoss);
 
-#endif  // FUSE_LOSS_SOFTLONE_LOSS_H
+#endif  // FUSE_LOSS_SCALED_LOSS_HPP_
