+++ fuse_loss/src/dcs_loss.cpp
@@ -32,8 +32,8 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  */
 #include <fuse_core/parameter.hpp>
-#include <fuse_loss/dcs_loss.h>
-#include <fuse_loss/loss_function.h>
+#include <fuse_loss/dcs_loss.hpp>
+#include <fuse_loss/loss_function.hpp>
 
 #include <pluginlib/class_list_macros.hpp>
