+++ fuse_loss/src/loss_function.cpp
@@ -31,7 +31,7 @@
  *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  *  POSSIBILITY OF SUCH DAMAGE.
  */
-#include <fuse_loss/loss_function.h>
+#include <fuse_loss/loss_function.hpp>
 
 #include <cmath>
 #include <limits>
