+++ fuse_loss/src/scaled_loss.cpp
@@ -31,7 +31,7 @@
  *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  *  POSSIBILITY OF SUCH DAMAGE.
  */
-#include <fuse_loss/scaled_loss.h>
+#include <fuse_loss/scaled_loss.hpp>
 
 #include <fuse_core/parameter.hpp>
 #include <pluginlib/class_list_macros.hpp>
