+++ fuse_loss/src/tukey_loss.cpp
@@ -31,7 +31,7 @@
  *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  *  POSSIBILITY OF SUCH DAMAGE.
  */
-#include <fuse_loss/tukey_loss.h>
+#include <fuse_loss/tukey_loss.hpp>
 
 #include <fuse_core/ceres_macros.hpp>
 #include <fuse_core/parameter.hpp>
