+++ fuse_loss/test/test_huber_loss.cpp
@@ -32,7 +32,7 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  */
 #include <fuse_core/serialization.hpp>
-#include <fuse_loss/huber_loss.h>
+#include <fuse_loss/huber_loss.hpp>
 
 #include <ceres/autodiff_cost_function.h>
 #include <ceres/problem.h>
