+++ fuse_loss/test/test_qwt_loss_plot.cpp
@@ -31,19 +31,19 @@
  *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  *  POSSIBILITY OF SUCH DAMAGE.
  */
-#include <fuse_loss/arctan_loss.h>
-#include <fuse_loss/cauchy_loss.h>
-#include <fuse_loss/dcs_loss.h>
-#include <fuse_loss/fair_loss.h>
-#include <fuse_loss/geman_mcclure_loss.h>
-#include <fuse_loss/huber_loss.h>
-#include <fuse_loss/softlone_loss.h>
-#include <fuse_loss/tolerant_loss.h>
-#include <fuse_loss/trivial_loss.h>
-#include <fuse_loss/tukey_loss.h>
-#include <fuse_loss/welsch_loss.h>
+#include <fuse_loss/arctan_loss.hpp>
+#include <fuse_loss/cauchy_loss.hpp>
+#include <fuse_loss/dcs_loss.hpp>
+#include <fuse_loss/fair_loss.hpp>
+#include <fuse_loss/geman_mcclure_loss.hpp>
+#include <fuse_loss/huber_loss.hpp>
+#include <fuse_loss/softlone_loss.hpp>
+#include <fuse_loss/tolerant_loss.hpp>
+#include <fuse_loss/trivial_loss.hpp>
+#include <fuse_loss/tukey_loss.hpp>
+#include <fuse_loss/welsch_loss.hpp>
 
-#include <fuse_loss/qwt_loss_plot.h>
+#include <fuse_loss/qwt_loss_plot.hpp>
 
 #include <gtest/gtest.h>
